requirements.txt
python-telegram-bot==20.7
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
pymongo==4.6.1
ffmpeg-python==0.2.0
python-dotenv==1.0.1 